+++ extensions/LemurProto/src/main/java/com/simsilica/lemur/OptionPanel.java
@@ -248,7 +248,6 @@ public class OptionPanel extends Panel {
      *  to provide different close behavior.
      */   
     public void close() {
-        GuiGlobals.getInstance().releaseCursorEnabled(this);
         if( hasEffect(EFFECT_CLOSE) ) {
             runEffect(EFFECT_CLOSE);
         } else if( getParent() instanceof Container ) {